Change setup.py to try without Cython and fallback on pure-python
The cython-generated .c file is included in sdist tarballs with
the MANIFEST.in file. (Make sure to run Cython befor sdist!)

If Cython is not installed, setup.py will fall back on using the
pre-generated .c file.

If even that fails, install anyway without the speedups.

This is inspired from the setup.py files for lxml and simplejson.