Make tests fail if an optional dependecy is missing
If lxml or the speedups are missing, the relevant tests are skipped.
(No need to show hundreds of failures.)

This adds one test for each optional dependecy, so that the
suite fails if something is missing, unless a specific enviroment
variable is set. (For when this is expected, like on PyPy.)